+++ index.js
@@ -1,11 +1,11 @@
 #!/usr/bin/env node
 
 const argv = require('yargs').argv,
-    server = require('./server');
+    server = require('./libs/server');
 const config = Object.assign({}, {
     path: './',
     port: 3000,
     https: true
-}, {...argv});
+}, argv);
 
 server(config);
